production/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.50.0
openai-agents==0.0.7
confluent-kafka==2.5.0
asyncpg==0.29.0
pgvector==0.3.0
twilio==9.0.0
google-api-python-client==2.140.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
structlog==24.4.0
sentence-transformers==3.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
locust==2.29.0
